modules/edge_api/requirements.txt
azure-iot-device~=2.7.0
Flask==0.10.1